update on routes

+++ database/migrations/2017_08_07_002940_create_sell_coins_table.php
@@ -19,7 +19,7 @@ class CreateSellCoinsTable extends Migration
             $table->string('wallet_id')->nullable();
             $table->string('buyer_email')->nullable();
             $table->bigInteger('buyer_phone')->nullable();
-            $table->bigInteger('amount_dollar')->nullable();
+            $table->float('amount_dollar')->nullable();
             $table->float('amount_btc')->nullable();
             $table->integer('rate')->nullable();
             $table->timestamps();
